feat: remove fake data and call the API on the server side
