--- src/app/(room)/room-type/fakeData.js
@@ -1,452 +1,0 @@
-const roomTypeFakeData = [
-  {
-    "_id": "65a4e27283315f6587b0cb44",
-    "name": "尊榮雙人房",
-    "description": "享受高級的住宿體驗，尊榮雙人房提供給您舒適寬敞的空間和精緻的裝潢。",
-    "imageUrl": "https://images.unsplash.com/photo-1616594039964-ae9021a400a0?q=80&w=1780&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-    "imageUrlList": [
-      "https://images.unsplash.com/photo-1696986681606-b156ccd761c5?q=80&w=1770&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "https://images.unsplash.com/photo-1616046619793-7e4badf3fe1f?q=80&w=1925&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "https://images.unsplash.com/photo-1638799869566-b17fa794c4de?q=80&w=1770&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "https://images.unsplash.com/photo-1566813326586-741374c2caa7?q=80&w=1872&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-    ],
-    "areaInfo": "24 坪",
-    "bedInfo": "1 張大床",
-    "maxPeople": 2,
-    "price": 10000,
-    "status": 1,
-    "facilityInfo": [
-      {
-        "title": "空調",
-        "isProvide": true
-      },
-      {
-        "title": "電視",
-        "isProvide": true
-      },
-      {
-        "title": "咖啡機",
-        "isProvide": true
-      },
-      {
-        "title": "熱水壺",
-        "isProvide": true
-      },
-      {
-        "title": "冰箱",
-        "isProvide": true
-      },
-      {
-        "title": "衣櫃",
-        "isProvide": true
-      },
-      {
-        "title": "除濕機",
-        "isProvide": false
-      },
-      {
-        "title": "保險箱",
-        "isProvide": true
-      },
-      {
-        "title": "電話",
-        "isProvide": true
-      },
-      {
-        "title": "沙發",
-        "isProvide": false
-      }
-    ],
-    "amenityInfo": [
-      {
-        "title": "浴巾",
-        "isProvide": true
-      },
-      {
-        "title": "浴袍",
-        "isProvide": true
-      },
-      {
-        "title": "洗髮乳",
-        "isProvide": true
-      },
-      {
-        "title": "護髮乳",
-        "isProvide": true
-      },
-      {
-        "title": "沐浴乳",
-        "isProvide": true
-      },
-      {
-        "title": "刮鬍刀",
-        "isProvide": true
-      },
-      {
-        "title": "牙刷",
-        "isProvide": true
-      },
-      {
-        "title": "吹風機",
-        "isProvide": true
-      },
-      {
-        "title": "香皂",
-        "isProvide": true
-      },
-      {
-        "title": "拖鞋",
-        "isProvide": true
-      }
-    ],
-    "createdAt": "2024-01-15T07:44:50.222Z",
-    "updatedAt": "2024-01-26T09:35:07.264Z",
-    "layoutInfo": [
-      {
-        "title": "市景",
-        "isProvide": false
-      },
-      {
-        "title": "獨立衛浴",
-        "isProvide": true
-      }
-    ]
-  },
-  {
-    "_id": "65a4e32683315f6587b0cb47",
-    "name": "市景雙人房",
-    "description": "市景雙人房擁有絕美的天際線景觀，讓您在舒適的環境中欣賞城市之美。",
-    "imageUrl": "https://images.unsplash.com/photo-1598928636135-d146006ff4be?q=80&w=1770&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-    "imageUrlList": [
-      "https://images.unsplash.com/photo-1696986681606-b156ccd761c5?q=80&w=1770&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "https://images.unsplash.com/photo-1616046619793-7e4badf3fe1f?q=80&w=1925&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "https://images.unsplash.com/photo-1638799869566-b17fa794c4de?q=80&w=1770&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "https://images.unsplash.com/photo-1566813326586-741374c2caa7?q=80&w=1872&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-    ],
-    "areaInfo": "28 坪",
-    "bedInfo": "1 張大床",
-    "maxPeople": 2,
-    "price": 12000,
-    "status": 1,
-    "facilityInfo": [
-      {
-        "title": "空調",
-        "isProvide": true
-      },
-      {
-        "title": "電視",
-        "isProvide": true
-      },
-      {
-        "title": "咖啡機",
-        "isProvide": true
-      },
-      {
-        "title": "熱水壺",
-        "isProvide": true
-      },
-      {
-        "title": "冰箱",
-        "isProvide": true
-      },
-      {
-        "title": "衣櫃",
-        "isProvide": true
-      },
-      {
-        "title": "除濕機",
-        "isProvide": true
-      },
-      {
-        "title": "保險箱",
-        "isProvide": true
-      },
-      {
-        "title": "電話",
-        "isProvide": true
-      },
-      {
-        "title": "沙發",
-        "isProvide": true
-      }
-    ],
-    "amenityInfo": [
-      {
-        "title": "浴巾",
-        "isProvide": true
-      },
-      {
-        "title": "浴袍",
-        "isProvide": true
-      },
-      {
-        "title": "洗髮乳",
-        "isProvide": true
-      },
-      {
-        "title": "護髮乳",
-        "isProvide": true
-      },
-      {
-        "title": "沐浴乳",
-        "isProvide": true
-      },
-      {
-        "title": "刮鬍刀",
-        "isProvide": true
-      },
-      {
-        "title": "牙刷",
-        "isProvide": true
-      },
-      {
-        "title": "吹風機",
-        "isProvide": true
-      },
-      {
-        "title": "香皂",
-        "isProvide": true
-      },
-      {
-        "title": "拖鞋",
-        "isProvide": true
-      }
-    ],
-    "createdAt": "2024-01-15T07:47:50.118Z",
-    "updatedAt": "2024-01-26T09:10:33.652Z",
-    "layoutInfo": [
-      {
-        "title": "市景",
-        "isProvide": true
-      }
-    ]
-  },
-  {
-    "_id": "65a4e39783315f6587b0cb4a",
-    "name": "玉璽雙床房",
-    "description": "與親朋摯友或閨蜜享受高級的住宿體驗。",
-    "imageUrl": "https://images.unsplash.com/photo-1618773928121-c32242e63f39?q=80&w=1770&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-    "imageUrlList": [
-      "https://images.unsplash.com/photo-1600566753151-384129cf4e3e?q=80&w=1770&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "https://raw.githubusercontent.com/hexschool/2022-web-layout-training/main/typescript-hotel/%E8%A1%8C%E5%8B%95%E7%89%88/room4-3.png",
-      "https://raw.githubusercontent.com/hexschool/2022-web-layout-training/main/typescript-hotel/%E8%A1%8C%E5%8B%95%E7%89%88/room4-4.png",
-      "https://raw.githubusercontent.com/hexschool/2022-web-layout-training/main/typescript-hotel/%E8%A1%8C%E5%8B%95%E7%89%88/room4-5.png"
-    ],
-    "areaInfo": "36 坪",
-    "bedInfo": "2 張大床",
-    "maxPeople": 4,
-    "price": 12800,
-    "status": 1,
-    "facilityInfo": [
-      {
-        "title": "空調",
-        "isProvide": true
-      },
-      {
-        "title": "電視",
-        "isProvide": true
-      },
-      {
-        "title": "咖啡機",
-        "isProvide": true
-      },
-      {
-        "title": "熱水壺",
-        "isProvide": true
-      },
-      {
-        "title": "冰箱",
-        "isProvide": true
-      },
-      {
-        "title": "衣櫃",
-        "isProvide": true
-      },
-      {
-        "title": "除濕機",
-        "isProvide": false
-      },
-      {
-        "title": "保險箱",
-        "isProvide": true
-      },
-      {
-        "title": "電話",
-        "isProvide": true
-      },
-      {
-        "title": "沙發",
-        "isProvide": true
-      }
-    ],
-    "amenityInfo": [
-      {
-        "title": "浴巾",
-        "isProvide": true
-      },
-      {
-        "title": "浴袍",
-        "isProvide": true
-      },
-      {
-        "title": "洗髮乳",
-        "isProvide": true
-      },
-      {
-        "title": "護髮乳",
-        "isProvide": true
-      },
-      {
-        "title": "沐浴乳",
-        "isProvide": true
-      },
-      {
-        "title": "刮鬍刀",
-        "isProvide": true
-      },
-      {
-        "title": "牙刷",
-        "isProvide": true
-      },
-      {
-        "title": "吹風機",
-        "isProvide": true
-      },
-      {
-        "title": "香皂",
-        "isProvide": true
-      },
-      {
-        "title": "拖鞋",
-        "isProvide": true
-      }
-    ],
-    "createdAt": "2024-01-15T07:49:43.502Z",
-    "updatedAt": "2024-01-26T09:56:33.901Z",
-    "layoutInfo": [
-      {
-        "title": "市景",
-        "isProvide": false
-      },
-      {
-        "title": "獨立衛浴",
-        "isProvide": true
-      }
-    ]
-  },
-  {
-    "_id": "65a4e40083315f6587b0cb4d",
-    "name": "玉尊總統房",
-    "description": "不僅有寬敞的空間，還有絕美的市景視野，帶給家庭最尊榮的待遇。",
-    "imageUrl": "https://images.unsplash.com/photo-1593987314040-8e0ac84ae724?q=80&w=1897&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-    "imageUrlList": [
-      "https://images.unsplash.com/photo-1618220321979-ecb3a975aa82?q=80&w=1932&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "https://images.unsplash.com/photo-1613013261832-a5d86dba9421?q=80&w=1771&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "https://images.unsplash.com/photo-1595514535116-d0401260e7cf?q=80&w=1770&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      "https://raw.githubusercontent.com/hexschool/2022-web-layout-training/main/typescript-hotel/%E8%A1%8C%E5%8B%95%E7%89%88/room5-5.png"
-    ],
-    "areaInfo": "48 坪",
-    "bedInfo": "2 張大床",
-    "maxPeople": 4,
-    "price": 20000,
-    "status": 1,
-    "facilityInfo": [
-      {
-        "title": "空調",
-        "isProvide": true
-      },
-      {
-        "title": "電視",
-        "isProvide": true
-      },
-      {
-        "title": "咖啡機",
-        "isProvide": true
-      },
-      {
-        "title": "熱水壺",
-        "isProvide": true
-      },
-      {
-        "title": "冰箱",
-        "isProvide": true
-      },
-      {
-        "title": "衣櫃",
-        "isProvide": true
-      },
-      {
-        "title": "除濕機",
-        "isProvide": false
-      },
-      {
-        "title": "保險箱",
-        "isProvide": true
-      },
-      {
-        "title": "電話",
-        "isProvide": true
-      },
-      {
-        "title": "沙發",
-        "isProvide": true
-      }
-    ],
-    "amenityInfo": [
-      {
-        "title": "浴巾",
-        "isProvide": true
-      },
-      {
-        "title": "浴袍",
-        "isProvide": true
-      },
-      {
-        "title": "洗髮乳",
-        "isProvide": true
-      },
-      {
-        "title": "護髮乳",
-        "isProvide": true
-      },
-      {
-        "title": "沐浴乳",
-        "isProvide": true
-      },
-      {
-        "title": "刮鬍刀",
-        "isProvide": true
-      },
-      {
-        "title": "牙刷",
-        "isProvide": true
-      },
-      {
-        "title": "吹風機",
-        "isProvide": true
-      },
-      {
-        "title": "香皂",
-        "isProvide": true
-      },
-      {
-        "title": "拖鞋",
-        "isProvide": true
-      }
-    ],
-    "createdAt": "2024-01-15T07:51:28.948Z",
-    "updatedAt": "2024-01-26T10:04:14.715Z",
-    "layoutInfo": [
-      {
-        "title": "市景",
-        "isProvide": false
-      },
-      {
-        "title": "獨立衛浴",
-        "isProvide": true
-      }
-    ]
-  }
-]
-
-export default roomTypeFakeData
